test(orderController): updated test for orders
the test of ordersController have now been created

resolveds partly #5
