+++ tests/Orders/controller/order.controller.getAll.spec.ts
@@ -1,0 +1,84 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { OrderController } from '../../../src/orders/order.controller';
+import { OrderService } from '../../../src/orders/order.service';
+import {
+  fakeOrderWithConfirmedStatus,
+  fakeOrderWithReceivedStatus,
+} from '../../testObjects';
+import { NotFoundException } from '@nestjs/common';
+import { NewOrderDto } from '../../../src/orders/dtos/new-order.dto';
+
+import { Ecommerce } from 'ckh-typings';
+
+describe('OrderController - Get All - The test files have been spilt up in multiple files for better maintainability', () => {
+  let controller: OrderController;
+  let fakeOrderService: Partial<OrderService>;
+  const OrderArray = [
+    fakeOrderWithReceivedStatus,
+    fakeOrderWithConfirmedStatus,
+  ];
+  const newOrderStatus = {
+    ...fakeOrderWithReceivedStatus,
+    orderStatus: Ecommerce.OrderStatus.PACKED,
+    updateLastChange: () => {},
+  };
+
+  beforeEach(async () => {
+    fakeOrderService = {
+      getAll: (page, limit) => {
+        return Promise.resolve([OrderArray, OrderArray.length]);
+      },
+      getOrderByStatus(status) {
+        const filteredOrder = OrderArray.filter(
+          (order) => order.orderStatus === status,
+        );
+        return Promise.resolve([filteredOrder, filteredOrder.length]);
+      },
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      createOne: (_newOrder: NewOrderDto) => {
+        return Promise.resolve(fakeOrderWithReceivedStatus);
+      },
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      getOne: (_id: number) => {
+        return Promise.resolve(fakeOrderWithReceivedStatus);
+      },
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      changeStatus: (_id: number, _newStatus: Ecommerce.OrderStatus) => {
+        return Promise.resolve(newOrderStatus);
+      },
+    };
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [{ provide: OrderService, useValue: fakeOrderService }],
+      controllers: [OrderController],
+    }).compile();
+
+    controller = module.get<OrderController>(OrderController);
+  });
+
+  describe('GetAll', () => {
+    test('There is no orders in the system, an error is thrown', async () => {
+      fakeOrderService.getAll = jest.fn().mockImplementation(() => {
+        throw new NotFoundException('There is no orders in the system');
+      });
+
+      expect.assertions(2);
+
+      try {
+        await controller.getAllOrders(1, 25);
+      } catch (error) {
+        expect(error).toBeInstanceOf(NotFoundException);
+        expect(error.message).toBe('There is no orders in the system');
+      }
+    });
+    test('The orders are return correctly when called with page and limit', async () => {
+      const ordersList = await controller.getAllOrders(1, 25);
+
+      expect.assertions(5);
+      expect(ordersList['orders'].length).toBe(2);
+      expect(ordersList.orders).toContain(fakeOrderWithReceivedStatus);
+      expect(ordersList.page).toBe(1);
+      expect(ordersList.limit).toBe(25);
+      expect(ordersList.totalCount).toBe(ordersList['orders'].length);
+    });
+  });
+});
